requirements.txt
uagents>=0.18.0
uagents-core>=0.2.0
openai>=1.40.0
fastapi>=0.110.0
uvicorn[standard]>=0.29.0
python-dotenv>=1.0.0
google-auth>=2.29.0
google-auth-oauthlib>=1.2.0
google-api-python-client>=2.120.0
pydantic>=2.6.0
httpx>=0.27.0
